backend/requirements.txt
asgiref==3.8.1
easy-thumbnails==2.10
gunicorn==23.0.0
Django==5.2.3
django-ckeditor==6.7.3
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
requests==2.32.4
singlemodeladmin==0.9
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
